gave RenderSpace's color func the generated value context
this made it possible to write `geneticPlantsRender`, in Shape.Plant,
since the color of a given record depends on the actual value generated,
rather than being wholly static. since the generated value is already
known when `color` is called (in Shape.pickSpace / Shape.pickSpaceUVs)
this is a trivial change, and mostly involves updating all the other
RenderSpace values.